votes/views//export: move export to own file and use gettext there
